it works fine without laziness..
